Some re-factory. Use spring to manage bean instead of JSF, use aop:scope-proxy to inject different scope of bean. Implement view scope. Update spring framework to 3.1.0.RELEASE

+++ mcs-web/src/com/mycallstation/web/converter/DurationConverter.java
@@ -13,7 +13,7 @@ import javax.faces.convert.Converter;
 import org.springframework.stereotype.Component;
 
 /**
- * @author wgao
+ * @author Wei Gao
  * 
  */
 @Component("durationConverter")
@@ -39,7 +39,7 @@ public class DurationConverter implements Converter {
 			int h = Integer.parseInt(m.group(1));
 			int mi = Integer.parseInt(m.group(2));
 			int s = Integer.parseInt(m.group(3));
-			long ret = ((long) (h * 3600 + mi * 60 + s)) * 1000L;
+			long ret = (h * 3600 + mi * 60 + s) * 1000L;
 			return ret;
 		}
 		return null;
